# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ATOM Sovereign Agent Runtime API
  description: |
    ATOM.0-alpha — Sovereign Recursive Agent Architecture.
    
    This API exposes the runtime surface for mission control, effect dispatch,
    capability management, and evidence retrieval. All mutation paths traverse
    the sovereign kernel boundary (authorize → commit).
    
    ## Authentication

    Every route except `GET /health` and `GET /ready` requires
    `Authorization: Bearer <token>`. The token is a provisioned operator
    secret (opaque, minimum 16 bytes), not a JWT: it proves the caller may
    talk to the daemon, never *who* they are. Missing or wrong credentials
    yield `401` with a `WWW-Authenticate: Bearer realm="atom"` challenge.
    The daemon refuses to start without a token unless `--no-auth` is passed
    explicitly (loopback development only). Auth is transport only — it never
    substitutes for capability grants or approvals.
    
    ## Idempotency
    All mutation endpoints accept `Idempotency-Key` header.
    Duplicate requests with the same key return the original response.
    
    ## Error Format
    Errors follow RFC 9457 (Problem Details):
    ```json
    {
      "type": "https://atom.dev/errors/kernel-denied",
      "title": "Kernel Authorization Denied",
      "status": 403,
      "detail": "Capability grant does not cover requested operation",
      "instance": "/missions/abc123/effects/def456"
    }
    ```
  version: 4.0.0-alpha
  license:
...
servers:
...
paths:
...
  /host/plan:
    post:
      tags: [host]
      summary: Declare a host mutation
      operationId: planHostOp
      description: |
        Declares a typed host operation as a durable effect intent and returns
        the effect digest an owner must approve (EFX-001).

        This endpoint reaches no executor and changes nothing on the host: a
        plan is a declaration, not authority. The capability grant must already
        cover the operation and resource, otherwise the plan is refused before
        it is recorded (deny-by-default).

        Requires the daemon to be started with `--host-root`; without it the
        host surface is disabled and every request is refused with 503.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostPlanRequest'
      responses:
        '201':
          description: Mutation declared; nothing has crossed the boundary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          description: Host mutation surface is disabled (no --host-root)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /host/commit:
    post:
      tags: [host]
      summary: Commit a declared host mutation
      operationId: commitHostOp
      description: |
        Redeems a durable approval grant for the plan's exact effect digest,
        re-validates authority and the resource witness at the boundary, issues
        a one-shot CommitPermit bound to this connector identity, and only then
        asks the privilege broker to admit the operation (KRN-002, EFX-004,
        AUT-003).

        The one-shot nonce is burned durably before the response returns, so the
        guarantee survives a restart. The approval is spent in the same step: an
        exact-effect approval authorises exactly one crossing.

        Refusals: no approval for this digest (400), resource witness drifted
        since planning (409), plan already committed (409), privilege boundary
        denied the crossing (409), host surface disabled (503).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostCommitRequest'
      responses:
        '200':
          description: The host was changed through the privilege boundary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostCommitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Witness drift, spent plan, or boundary denial
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          description: Host mutation surface is disabled (no --host-root)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Operator bearer token (`Authorization: Bearer <token>`). Required on
        every operation except `GET /health` and `GET /ready`. Transport only:
        it never substitutes for capability grants or approvals.
...
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    HostOp:
      description: |
        A single host-administration operation. The set is closed and internally
        tagged on `op`: there is no "run arbitrary command" variant, and an
        unknown tag is refused at deserialization (deny-by-default).
      oneOf:
        - type: object
          additionalProperties: false
          required: [op, path, contents]
          properties:
            op:
              const: write_file
            path:
              type: string
              description: Absolute path, resolved inside the sandbox root.
            contents:
              type: string
        - type: object
          additionalProperties: false
          required: [op, path]
          properties:
            op:
              const: remove_file
            path:
              type: string
        - type: object
          additionalProperties: false
          required: [op, program, args]
          properties:
            op:
              const: spawn_process
            program:
              type: string
              description: Absolute path inside the sandbox; never resolved through PATH.
            args:
              type: array
              items:
                type: string
        - type: object
          additionalProperties: false
          required: [op, interface, allow_cidr]
          properties:
            op:
              const: configure_network
            interface:
              type: string
            allow_cidr:
              type: string
              description: IPv4 CIDR. Refused by the filesystem sandbox executor.
      discriminator:
        propertyName: op
    ResourceWitness:
      type: object
      description: |
        The observed version of a resource. `kind` is `sha256` for an existing
        file or `absent` when the resource does not exist — an absent resource is
        a witness too, so creating a file somebody else created is drift.
      required: [kind, resource_id, value]
      properties:
        kind:
          type: string
        resource_id:
          type: string
        value:
          type: string
    HostPlanRequest:
      type: object
      additionalProperties: false
      required: [mission_id, grant_id, op]
      properties:
        mission_id:
          type: string
        grant_id:
          type: string
          description: The capability grant the authority must come from.
        op:
          $ref: '#/components/schemas/HostOp'
    HostPlanResponse:
      type: object
      required:
        - plan_id
        - mission_id
        - grant_id
        - operation
        - resource_type
        - resource_id
        - effect_digest
        - planned_witness
        - state
      properties:
        plan_id:
          type: string
        mission_id:
          type: string
        grant_id:
          type: string
        operation:
          type: string
          enum: [write, delete, spawn, configure]
        resource_type:
          type: string
          enum: [file, process, network]
        resource_id:
          type: string
        effect_digest:
          type: string
          description: The identity an exact-effect approval must name.
        planned_witness:
          $ref: '#/components/schemas/ResourceWitness'
        state:
          const: PLANNED
    HostCommitRequest:
      type: object
      additionalProperties: false
      required: [plan_id]
      properties:
        plan_id:
          type: string
    HostCommitResponse:
      type: object
      required:
        - plan_id
        - permit_id
        - one_shot_nonce
        - approval_id
        - effect_digest
        - outcome
        - observation_id
        - state
      properties:
        plan_id:
          type: string
        permit_id:
          type: string
        one_shot_nonce:
          type: string
          description: Burned durably before this response returned.
        approval_id:
          type: string
        effect_digest:
          type: string
        outcome:
          type: string
          description: The executor's own account of what it did.
        observation_id:
          type: string
          description: The sealed evidence entry, tainted HOST_MUTATION_COMMITTED.
        state:
          const: COMMITTED
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'